polish examples; add jax as dependency; better error in sample function
