+++ examples/model.py
@@ -60,20 +60,4 @@ hists = {
 }
 
 observation = jnp.array([37])
-
-nll = evm.loss.PoissonNLL()
-
-
-@eqx.filter_jit
-def loss(model, hists, observation):
-    expectations = model(hists)
-    constraints = evm.loss.get_param_constraints(model)
-    return nll(
-        expectation=evm.util.sum_leaves(expectations),
-        observation=observation,
-        constraint=evm.util.sum_leaves(constraints),
-    )
-
-
-loss_val = loss(model, hists, observation)
-grads = eqx.filter_grad(loss)(model, hists, observation)
+expectations = model(hists)
