+++ examples/nll_fit.py
@@ -1,0 +1,39 @@
+import equinox as eqx
+import jax.numpy as jnp
+import optax
+from model import hists, model, observation
+
+import evermore as evm
+
+optim = optax.sgd(learning_rate=1e-2)
+opt_state = optim.init(eqx.filter(model, eqx.is_inexact_array))
+
+nll = evm.loss.PoissonNLL()
+
+
+@eqx.filter_jit
+def loss(model, hists, observation):
+    expectations = model(hists)
+    constraints = evm.loss.get_param_constraints(model)
+    loss_val = nll(
+        expectation=evm.util.sum_leaves(expectations),
+        observation=observation,
+    )
+    # add constraint
+    loss_val += evm.util.sum_leaves(constraints)
+    return -jnp.sum(loss_val)
+
+
+@eqx.filter_jit
+def make_step(model, opt_state, events, observation):
+    # differentiate full analysis
+    grads = eqx.filter_grad(loss)(model, events, observation)
+    updates, opt_state = optim.update(grads, opt_state)
+    # apply nuisance parameter and DNN weight updates
+    model = eqx.apply_updates(model, updates)
+    return model, opt_state
+
+
+# minimize model with 1000 steps
+for _ in range(1000):
+    model, opt_state = make_step(model, opt_state, hists, observation)
