+++ pyproject.toml
@@ -26,6 +26,8 @@ classifiers = [
 ]
 dynamic = ["version"] # version is set in src/evermore/__init__.py
 dependencies = [
+    "jax",
+    "jaxtyping",
     "equinox>=0.10.6", # eqx.field
 ]
